requirements.txt
pandas 
numpy
sqlalchemy
#psycopg2
psycopg2-binary
PyYAML
tabula-py
JPype1